mb/google/fatcat: Adjust EC host command range for fatcat_ish variant
Adjusts the EC host command range for the fatcat_ish variant
to 0x800-0x807 & 0x200-0x20f.

This change is necessary because the microchip EC used on the Fatcat
board has a smaller host command range than the ITE/Nuvoton ECs used
on other Fatcat variants.

without this patch:

    [SPEW ]  LPC: Trying to open IO window from 800 size 8
    [ERROR]  LPC: Cannot open IO window: 800 size 8
    [ERROR]  No more IO windows

with this patch:
    [SPEW ]  LPC: Trying to open IO window from 800 size 8

BUG=b:370984186
TEST=Able to build fatcat/fatcat_ish w/o any error.

Change-Id: I0d726d60d2a15d2dfaff35f570de479fdc6d15aa
Signed-off-by: Jayvik Desai <jayvik@google.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/85221
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Pranava Y N <pranavayn@google.com>
Reviewed-by: Subrata Banik <subratabanik@google.com>
Reviewed-by: Kapil Porwal <kapilporwal@google.com>