Documentation: Fix wrong link to commit message guidelines
In the Gerrit guidelines there is an old link pointing to the retired
coreboot wiki (https://www.coreboot.org/Git#Commit_messages) when the
commit message guidelines are referenced. Indeed this section was never
ported over to the new documentation and is missing.

This commit rewrites this guidelines and adds them as a new section
based on what was in the wiki and updates the link accordingly.

Change-Id: I1cd2b13da6fe59697d677c7350d73eda5d486544
Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/85915
Reviewed-by: Maximilian Brune <maximilian.brune@9elements.com>
Reviewed-by: Felix Singer <service+coreboot-gerrit@felixsinger.de>
Reviewed-by: Matt DeVillier <matt.devillier@gmail.com>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>