mb/google/fatcat/var/fatcat: Drop UFC/WFC GPIO programming
Skip UFC/WFC GPIO programming for power-on and clock configuration.

Clock configuration is now handled by native-function in ramstage,
and there is no need to power-on the camera module early in the boot
phase. Doing so resulted in the privacy LED being turned on during the
entire boot process, which is unnecessary.

BUG=b:381044394
TEST=No privacy LED blinking seen while booting google/fatcat.

Change-Id: Iae984a2ab6f797af450166c90f4a2c6d3e0e1caa
Signed-off-by: Subrata Banik <subratabanik@google.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/85955
Reviewed-by: Eric Lai <ericllai@google.com>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>