mb/asrock/z87_extreme4: Select correct Kconfig for ASM1061
Commit fee8bcbcfb55 ("drivers/asmedia: Add code to enable AHCI for
ASM1061") renamed the ASMEDIA_ASPM_BLACKLIST Kconfig to ASMEDIA_ASM_1061
since the former was only ever used for that card.

However, commit fb2462053497 ("mb/asrock: Add Z87 Extreme4 (Haswell)")
was merged first and used the old Kconfig, so the tree broke when the
AHCH patch was merged. Change the Kconfig to fix the tree.

Change-Id: If16455c43e583e7edbf7914db87346385d811190
Signed-off-by: Nicholas Chin <nic.c3.14@gmail.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/86056
Reviewed-by: Jan Philipp Groß <jeangrande@mailbox.org>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Elyes Haouas <ehaouas@noos.fr>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>