mb/google/fatcat: Move Finger Print Sensor (FPS) from GSPI0A to GSPI0
This moves the FPS device from GSPI0A to GSPI0 to align with the
hardware design dated Jan'25.

The FPS device was initially placed on GSPI0A, which was incorrect. This
commit rectifies the configuration by moving it to the correct GSPI0
interface.

This change ensures that the CRFP device is correctly connected and
functions as expected.

BUG=b:377595986
TEST=Able to build and boot google/fatcat.

Change-Id: I3996f1a054204689ad733c650b6f71f1482c0b22
Signed-off-by: Subrata Banik <subratabanik@google.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/86143
Reviewed-by: Paul Menzel <paulepanter@mailbox.org>
Reviewed-by: Pranava Y N <pranavayn@google.com>
Reviewed-by: Kapil Porwal <kapilporwal@google.com>
Reviewed-by: Jayvik Desai <jayvik@google.com>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>